requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.11.10
botocore==1.14.10
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
fanstatic==1.1
Flask==1.1.1
Flask-WTF==0.14.3
gensim==3.8.1
idna==2.8
IMAPClient==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
networkx==2.4
nltk==3.4.5
numpy==1.18.1
olefile==0.46
pandas==1.0.0
pdfminer.six==20181108
Pillow==7.0.0
pycryptodome==3.9.6
pyparsing==2.4.6
PyPDF2==1.26.0
pyspellchecker==0.5.3
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
requests==2.22.0
s3transfer==0.3.2
scikit-learn==0.22.1
scipy==1.4.1
shutilwhich==1.1.0
six==1.12.0
sklearn==0.0
smart-open==1.9.0
sortedcontainers==2.1.0
soupsieve==1.9.5
textract==1.6.3
tzlocal==1.5.1
urllib3==1.25.8
WebOb==1.8.6
Werkzeug==1.0.0
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.7
